# server/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/settings": {
      "get": {
        "description": "后台设置页初始化用。",
        "operationId": "AdminSettingsController_getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "完整配置",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteSettingsValuesDto"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "读取全部配置（含说明）",
        "tags": [
          "站点设置 · 管理员"
        ]
      },
      "patch": {
        "description": "只接受白名单内的键，其余键静默忽略（不是报错，方便前端整表单提交）。更新后缓存立即失效。",
        "operationId": "AdminSettingsController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新结果",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateSettingsResultDto"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "批量更新配置",
        "tags": [
          "站点设置 · 管理员"
        ]
      }
    },
    "/api/v1/admin/settings/sensitive-words": {
      "get": {
        "description": "从 `site_settings.sensitive_words` 解析（支持逗号、中文逗号、换行分隔）。",
        "operationId": "AdminSettingsController_getSensitiveWords",
        "parameters": [],
        "responses": {
          "200": {
            "description": "敏感词数组",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "读取敏感词列表",
        "tags": [
          "站点设置 · 管理员"
        ]
      }
    },
...
    "/api/v1/admin/categories": {
      "post": {
        "description": "仅管理员。作者只能在前台选用分类，不能创建（FR-3.1）。",
        "operationId": "AdminCategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "分类名或 slug 已存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "新建分类",
        "tags": [
          "分类 · 管理员"
        ]
      }
    },
    "/api/v1/admin/categories/{id}": {
      "patch": {
        "operationId": "AdminCategoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "int64",
              "example": "1",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "修改分类",
        "tags": [
          "分类 · 管理员"
        ]
      },
      "delete": {
        "description": "外键为 `ON DELETE SET NULL`：**分类下的文章不会被删除**，只会变为「未分类」。响应里会告知受影响的文章数。",
        "operationId": "AdminCategoriesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功并返回受影响文章数"
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除分类",
        "tags": [
          "分类 · 管理员"
        ]
      }
    },
    "/api/v1/admin/tags": {
      "post": {
        "description": "作者在发文时也能自动创建标签，此接口供后台手工补建。",
        "operationId": "AdminTagsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTagDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "新建标签",
        "tags": [
          "标签 · 管理员"
        ]
      }
    },
    "/api/v1/admin/tags/merge": {
      "post": {
        "description": "把多个标签合并为一个，典型场景是清理「JS / javascript / Javascript」这类同义标签。合并后源标签被删除，其文章改挂到目标标签。",
        "operationId": "AdminTagsController_merge",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeTagsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "合并结果"
          },
          "400": {
            "description": "源标签与目标标签相同或为空",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "合并标签",
        "tags": [
          "标签 · 管理员"
        ]
      }
    },
    "/api/v1/admin/tags/{id}": {
      "delete": {
        "description": "关联的 `article_tags` 记录由外键级联清理，文章本身不受影响。",
        "operationId": "AdminTagsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResultDto"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除标签",
        "tags": [
          "标签 · 管理员"
        ]
      }
    },
...
    "/api/v1/me/upload/image": {
      "post": {
        "description": "用于文章封面与正文配图。**三道校验**：扩展名白名单、MIME 白名单、文件头魔数比对。\n第三道是关键 —— 前两道都能被伪造（改扩展名、改 Content-Type 即可），只有真实字节不会骗人。\n\n文件按 `年/月` 分目录存放，文件名使用「时间戳 + 内容哈希 + 随机段」，不保留原始文件名。\n响应里会带上图片宽高，前端可用它预留占位、避免图片加载时的布局抖动。",
        "operationId": "UploadController_uploadImage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "图片文件",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "jpg / png / gif / webp，默认上限 5MB"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "上传成功，返回可访问的 URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadedImageDto"
                }
              }
            }
          },
          "400": {
            "description": "类型不支持、内容与声明不符或超出大小限制",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "上传图片",
        "tags": [
          "上传"
        ]
      }
    },
...
  },
  "info": {
    "title": "博客平台 API",
    "description": "多作者博客平台后端接口。\n\n## 认证方式\n登录后拿到 `accessToken`，在请求头带上 `Authorization: Bearer <token>`。\nAccess Token 默认 15 分钟过期，用 `POST /auth/refresh` 换成新的令牌对。\n\n## 关于 ID\n所有主键在数据库里是 `BIGINT`，接口统一以**字符串**收发 ——\nJS 的 `Number` 在 2^53 以上会静默丢精度，这是比报错更难排查的问题。\n\n## 关于分页\n列表接口统一返回 `{ items, total, page, pageSize, totalPages }`。\n\n## 关于权限\n角色两级：`admin`（全站）与 `author`（只能操作自己的内容）。\n另外账号需要处于 `active` 状态才能执行写操作 —— 注册后是 `pending`，需管理员审核激活。",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "填入 accessToken（不含 Bearer 前缀）"
      }
    },
    "schemas": {
...
      "SiteSettingsValuesDto": {
        "type": "object",
        "properties": {
          "values": {
            "type": "object",
            "description": "配置键值对",
            "example": {
              "site_name": "我的博客",
              "per_page": "12"
            }
          },
          "descriptions": {
            "type": "object",
            "description": "配置项说明，键与 values 对应"
          }
        },
        "required": [
          "values",
          "descriptions"
        ]
      },
      "UpdateSettingsDto": {
        "type": "object",
        "properties": {
          "values": {
            "type": "object",
            "description": "要更新的配置键值对。只接受白名单内的键（站点信息、每页条数、注册与评论开关、敏感词、社交链接），其余会被忽略",
            "example": {
              "site_name": "我的博客",
              "comment_need_approve": "false"
            }
          }
        },
        "required": [
          "values"
        ]
      },
      "UpdateSettingsResultDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "是否成功",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "结果说明"
          },
          "updated": {
            "description": "实际被更新的配置键",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "success",
          "message",
          "updated"
        ]
      },
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "description": "HTTP 状态码",
            "example": 400
          },
          "error": {
            "type": "string",
            "description": "错误简述",
            "example": "Bad Request"
          },
          "message": {
            "description": "错误详情；校验失败时为字符串数组",
            "example": [
              "password 长度不能少于 8 位"
            ],
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "path": {
            "type": "string",
            "description": "请求路径",
            "example": "/api/v1/articles"
          },
          "timestamp": {
            "type": "string",
            "description": "发生时间（ISO 8601）",
            "example": "2026-09-17T09:30:00.000Z"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "path",
          "timestamp"
        ]
      },
...
      "OperationResultDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "是否成功",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "附带说明",
            "example": "已移入回收站"
          }
        },
        "required": [
          "success"
        ]
      },
...
      "CategoryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "分类 ID",
            "example": "1"
          },
          "name": {
            "type": "string",
            "description": "分类名",
            "example": "后端"
          },
          "slug": {
            "type": "string",
            "description": "URL 标识",
            "example": "backend"
          },
          "description": {
            "type": "string",
            "description": "描述",
            "nullable": true
          },
          "sortOrder": {
            "type": "number",
            "description": "排序值",
            "example": 0
          },
          "articleCount": {
            "type": "number",
            "description": "该分类下已发布的文章数",
            "example": 24
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "sortOrder",
          "articleCount"
        ]
      },
      "TagResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "标签 ID",
            "example": "1"
          },
          "name": {
            "type": "string",
            "description": "标签名",
            "example": "PostgreSQL"
          },
          "slug": {
            "type": "string",
            "description": "URL 标识",
            "example": "postgresql"
          },
          "articleCount": {
            "type": "number",
            "description": "关联的文章数",
            "example": 8
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "articleCount"
        ]
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "分类名",
            "example": "后端"
          },
          "slug": {
            "type": "string",
            "description": "URL 标识；留空则根据名称生成。必须是全小写字母/数字/短横线",
            "example": "backend"
          },
          "description": {
            "type": "string",
            "description": "分类描述",
            "example": "服务端、数据库、架构相关"
          },
          "sortOrder": {
            "type": "number",
            "description": "排序值，越小越靠前",
            "default": 0
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "分类名",
            "example": "后端"
          },
          "slug": {
            "type": "string",
            "description": "URL 标识；留空则根据名称生成。必须是全小写字母/数字/短横线",
            "example": "backend"
          },
          "description": {
            "type": "string",
            "description": "分类描述",
            "example": "服务端、数据库、架构相关"
          },
          "sortOrder": {
            "type": "number",
            "description": "排序值，越小越靠前",
            "default": 0
          }
        }
      },
      "CreateTagDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "标签名",
            "example": "PostgreSQL"
          },
          "slug": {
            "type": "string",
            "description": "URL 标识；留空自动生成",
            "example": "postgresql"
          }
        },
        "required": [
          "name"
        ]
      },
      "MergeTagsDto": {
        "type": "object",
        "properties": {
          "sourceTagIds": {
            "description": "被合并掉的标签 ID 列表（合并后这些标签会被删除，其文章改挂到目标标签）",
            "example": [
              "3",
              "7"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "targetTagId": {
            "type": "string",
            "description": "保留的目标标签 ID",
            "example": "1"
          }
        },
        "required": [
          "sourceTagIds",
          "targetTagId"
        ]
      },
...
      "UploadedImageDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "description": "可公开访问的图片地址，直接用于 Markdown 图片语法或封面字段",
            "example": "http://localhost:3000/uploads/2026/09/1726xxxxxxxx-a1b2c3d4.webp"
          },
          "path": {
            "type": "string",
            "description": "服务器上的绝对路径，用于排查与备份脚本"
          },
          "filename": {
            "type": "string",
            "description": "生成的文件名（时间戳 + 内容哈希 + 随机段，不保留原始名）",
            "example": "1726xxxxxxxx-a1b2c3d4.webp"
          },
          "size": {
            "type": "number",
            "description": "文件字节数",
            "example": 128456
          },
          "mimeType": {
            "type": "string",
            "description": "经文件头校验后的真实 MIME 类型",
            "example": "image/webp"
          },
          "width": {
            "type": "number",
            "description": "图片宽度（px）。可用于预留占位、避免加载时布局抖动",
            "example": 1280
          },
          "height": {
            "type": "number",
            "description": "图片高度（px）",
            "example": 720
          }
        },
        "required": [
          "url",
          "path",
          "filename",
          "size",
          "mimeType"
        ]
      },
...
    }
  }
}